feat: better low qubit counts
